==> data_cleaning/01_data_inspection+cleaning.ipynb <==
# Data Inspection
Author(s): Tomaso Stefanizzi

Description:
The dataset used for this project contains a historical data about brazilian flights between 2000 until 2021.
You can find the source 
[here](https://www.kaggle.com/datasets/mayconfelipemota/brazilian-data-flights).

### Imports

#import libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os as os

### read data

dest_path = 'dataset/DW_ARPT_DEST.csv'
orig_path = 'dataset/DW_ARPT_ORIGEM.csv'
company_path = 'dataset/DW_EMPRESA.csv'
equip_path = 'dataset/DW_EQPT.csv'
line_path = 'dataset/DW_TIPO_LINHA.csv'
#voos_path = 'dataset/DW_VOO.csv'

#read the data
dest = pd.read_csv(dest_path, encoding='latin1')
orig = pd.read_csv(orig_path, encoding='latin1')
company = pd.read_csv(company_path, encoding='latin1')
equip = pd.read_csv(equip_path, encoding='latin1')
line = pd.read_csv(line_path, encoding='latin1')
#voos = pd.read_csv(voos_path)

# Tables inspection

## Airport Destination

dest.head()

dest.info()

#### let's see if there are duplicates in the id (which i assume to be ```id_aerodromo_destino```)

dest['id_aerodromo_destino'].value_counts()


and indeed there are :D let's keep just one (possibly with all the informations)

dest[dest['id_aerodromo_destino']==34]

#remove duplicates
dest = dest.drop_duplicates()

dest.info()

dupl = dest['id_aerodromo_destino'].value_counts() 
dupl[dupl>1]

#### id 802:

dest[dest['id_aerodromo_destino']==802]

#drop line with index 0
dest = dest.drop([111,1244])


#### id 17:


dest[dest['id_aerodromo_destino']==17]

I kept the correct code

dest = dest.drop(399)

#### id 34:

dest[dest['id_aerodromo_destino']==34]

dest = dest.drop(96)

#### id 10:

dest[dest['id_aerodromo_destino']==10]

dest = dest.drop(639)

#### id 1292:

dest[dest['id_aerodromo_destino']==1292]

dest = dest.drop(1009)

#### id 1479:

dest[dest['id_aerodromo_destino']==1479]

dest = dest.drop(996)

#### id 208:

dest[dest['id_aerodromo_destino']==208]

dest = dest.drop(698)

#### id 21243:

also here i kept the correct code

dest[dest['id_aerodromo_destino']==21243]

dest = dest.drop(976)

#### id 338:

and so here

dest[dest['id_aerodromo_destino']==338]

dest = dest.drop(652)

## Origin

orig.head()

orig.info()

#### also here, let's see if the ids are unique

dupl = orig['id_aerodromo_origem'].value_counts()
dupl[dupl>1]

there are too many duplicates. for this reason, I want it to be consistent with the ones from the departure, and the others i'll just keep the same values for the ones in departures, and then for the others I'll try another strategy.

-   802      3
-   17       2
-   34       2
-   10       2
-   1292     2
-   1479     2
-   208      2
-   21243    2
-   338      2

#### strategy 1:

##### id 802:

orig[orig['id_aerodromo_origem']==802]

orig = orig.drop([119,1211,1482])

##### id 17:

orig[orig['id_aerodromo_origem']==17]

orig = orig.drop([395, 1246, 1393])   

##### id 34:

orig[orig['id_aerodromo_origem']==34]

orig = orig.drop([53, 1134, 1266, 1427])

##### id 10:

orig[orig['id_aerodromo_origem']==10]

orig = orig.drop([632, 1129, 1240,1418])

##### id 1292:

orig[orig['id_aerodromo_origem']==1292]

orig = orig.drop([982, 1135, 1256, 1409])

##### id 1479:

orig[orig['id_aerodromo_origem']==1479]

orig = orig.drop([965, 113, 1323])

##### id 208:

orig[orig['id_aerodromo_origem']==208]

orig = orig.drop([692, 1173, 1475])

##### id 21243:

orig[orig['id_aerodromo_origem']==21243]

orig = orig.drop([941, 944, 1020, 1364])

##### id 338:

orig[orig['id_aerodromo_origem']==338]

orig = orig.drop(648)

#### Strategy 2

using the mode() function in combination with groupby(), we can find the most frequent value in a series. for each group, we want to find the most frequent value in each column. the result will be a new DataFrame with the most popular attributes for each id

df_grouped = orig.groupby('id_aerodromo_origem', as_index=False )
# Use mode() to get the most frequent values in each group
most_popular = df_grouped.agg(lambda x: x.value_counts().index[0] if not x.value_counts().empty else np.nan)

most_popular
orig = most_popular

## Companies

company.head()

company.info()

#### as before, I'm checking if there are duplicates. if there are, then use the same approach of strategy 2 for origin dataset

dupl = company['id_empresa'].value_counts()
dupl[dupl>1]

since there are a lot of duplicates, I proceed as before

# keep as before the most frequent values
df_grouped = company.groupby('id_empresa', as_index=False )
# Use mode() to get the most frequent values in each group
most_popular = df_grouped.agg(lambda x: x.value_counts().index[0] if not x.value_counts().empty else np.nan)

most_popular
company = most_popular

company.head()

company.info()

## Equipements

equip.head()

equip.info()

equip['id_equipamento'].value_counts()

# keep as before the most frequent values
df_grouped = equip.groupby('id_equipamento', as_index=False )
# Use mode() to get the most frequent values in each group
most_popular = df_grouped.agg(lambda x: x.value_counts().index[0] if not x.value_counts().empty else np.nan)
equip = most_popular

equip.head()

equip.info()

## Route Type

line.head()

line.info()

line['id_tipo_linha'].value_counts()

# keep as before the most frequent values
df_grouped = line.groupby('id_tipo_linha', as_index=False )
# Use mode() to get the most frequent values in each group
most_popular = df_grouped.agg(lambda x: x.value_counts().index[0] if not x.value_counts().empty else np.nan)
line = most_popular

line.head()

line.info()

## Flights

voos_path = 'dataset/DW_VOOS.csv'
voos = pd.read_csv(voos_path, encoding='latin1')

voos.head()

voos.columns

voos.info()

dupl = voos['id_basica'].value_counts()
dupl[dupl>1]

no duplicates here:D

## first round of cleaned data export

#export data
dest.to_csv('dataset/modified_data/DW_ARPT_DEST.csv', index=False)
orig.to_csv('dataset/modified_data/DW_ARPT_ORIGEM.csv', index=False)
company.to_csv('dataset/modified_data/DW_EMPRESA.csv', index=False)
equip.to_csv('dataset/modified_data/DW_EQPT.csv', index=False)
line.to_csv('dataset/modified_data/DW_TIPO_LINHA.csv', index=False)
