==> data_cleaning/02_data_cleaning.ipynb <==
# Data Inspection
Author(s): Tomaso Stefanizzi

Description:
In this notebook I'm cleaning the data + merging two tables

## Imports

#import libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os as os

### read data

dest_path = 'dataset/modified_data/DW_ARPT_DEST.csv'
orig_path = 'dataset/modified_data/DW_ARPT_ORIGEM.csv'
company_path = 'dataset/modified_data/DW_EMPRESA.csv'
equip_path = 'dataset/modified_data/DW_EQPT.csv'
line_path = 'dataset/modified_data/DW_TIPO_LINHA.csv'
#voos_path = 'dataset/DW_VOO.csv'

#read the data
dest = pd.read_csv(dest_path)
orig = pd.read_csv(orig_path)
company = pd.read_csv(company_path)
equip = pd.read_csv(equip_path)
line = pd.read_csv(line_path)
#voos = pd.read_csv(voos_path)

### dest

dest.head()

# drop the columns that are not useful
dest = dest.drop(['sg_uf_destino', 'nm_regiao_destino', 'nr_escala_destino'], axis=1)

dest.info()

# drop the null values in the dataset
dest = dest.dropna()

dest.info()

dest.drop(['DT_CARGA_DW'], axis=1, inplace=True)

### orig

orig.head()

#drop the columns that are not useful
orig = orig.drop(['sg_uf_origem', 'nm_regiao_origem'], axis=1)

orig.info()

#drop the null values in the dataset
orig = orig.dropna()

orig.info()

orig.drop(['DT_CARGA_DW'], axis=1, inplace=True)

### company

company.head()

company.info()

company = company.dropna()

company.info()

company.drop(['DT_CARGA_DW'], axis=1, inplace=True)

### equip

equip.head()

equip.info()

#drop the null values in the dataset
equip = equip.dropna()

equip.info()

equip.drop(['DT_CARGA_DW'], axis=1, inplace=True)

### line

line

line = line.drop([0])
line.info()

line.drop(['DT_CARGA_DW'], axis=1, inplace=True)

## And now the big guy

voos_path = 'dataset/DW_VOOS.csv'
voos = pd.read_csv(voos_path, encoding='latin1')

voos.head()

voos.columns

voos.info(verbose = True)

### what now?
I want to keep just informations that are available also in the other tables. in order to do so, I follow this logic:

keep flights with: 
-   cleaned id_empresa
-   cleaned id_equipamento
-   cleaned id_aerodromo_origem
-   cleaned id_aerodromo_destino
-   cleaned id_tipo_linea

after this, check the shape of the table and decide what to do

company_ids = company['id_empresa'].tolist()

voos = voos[voos['id_empresa'].isin(company_ids)]

voos.shape

equip_ids = equip['id_equipamento'].tolist()
voos = voos[voos['id_equipamento'].isin(equip_ids)]

voos.shape

orig_ids = orig['id_aerodromo_origem'].tolist()
voos = voos[voos['id_aerodromo_origem'].isin(orig_ids)]

voos.shape

dest_ids = dest['id_aerodromo_destino'].tolist()
voos = voos[voos['id_aerodromo_destino'].isin(dest_ids)]

voos.shape

line_ids = line['id_tipo_linha'].tolist()
voos = voos[voos['id_tipo_linha'].isin(line_ids)]

voos.shape

voos.info()

voos.dropna(inplace=True)
voos.shape

##### ok **maybe** 15,4 M of flights are a bit too much :/ I'm going to find a way to diminish it substantially

voos.dt_partida_real.unique()

# keep the flights where date of departure is after 2018
flights = voos.copy()
flights = flights[flights['dt_partida_real'] > '2018-01-01']
flights.shape

##### much better, but still 2,1 M is still a lot... maybe i'll just keep the flights for the last year (2020)

flights2 = flights.copy()
flights2 = flights2[flights2['dt_partida_real'] > '2020-01-01']

flights2.shape

##### ok 300K is more managable. Now, I have to make sure that the other tables just keep the useful informations that I have in the flights2 table

print(f'before: line: {line.shape}, orig: {orig.shape}, dest: {dest.shape}, equip: {equip.shape}, company: {company.shape}')

line_id = flights2['id_tipo_linha'].tolist()
line = line[line['id_tipo_linha'].isin(line_id)]

dep_id = flights2['id_aerodromo_origem'].tolist()
orig = orig[orig['id_aerodromo_origem'].isin(dep_id)]

dest_id = flights2['id_aerodromo_destino'].tolist()
dest = dest[dest['id_aerodromo_destino'].isin(dest_id)]

equip_id = flights2['id_equipamento'].tolist()
equip = equip[equip['id_equipamento'].isin(equip_id)]

company_id = flights2['id_empresa'].tolist()
company = company[company['id_empresa'].isin(company_id)]

print(f'after: line: {line.shape}, orig: {orig.shape}, dest: {dest.shape}, equip: {equip.shape}, company: {company.shape}')



### save the final data

#save the data
dest.to_csv('dataset/processed_data/destination.csv', index=False)
orig.to_csv('dataset/processed_data/origin.csv', index=False)
company.to_csv('dataset/processed_data/company.csv', index=False)
equip.to_csv('dataset/processed_data/equipment.csv', index=False)
line.to_csv('dataset/processed_ata/line.csv', index=False)
flights2.to_csv('dataset/processed_data/flights.csv', index=False)
