==> data_cleaning/04_save_files_correctly.ipynb <==
# Final Data Inspection
Author(s): Tomaso Stefanizzi

Description:
In this notebook I'm just setting the csv to have the a different separation character

import pandas as pd

#### read files

airport_path = 'dataset/final_data/airports.csv'
airline_path = 'dataset/processed_data/company.csv'
route_path = 'dataset/processed_data/line.csv'
flight_path = 'dataset/final_data/flights.csv'
equipment_path = 'dataset/processed_data/equipment.csv'

airport = pd.read_csv(airport_path)
airline = pd.read_csv(airline_path)
route = pd.read_csv(route_path)
flight = pd.read_csv(flight_path)
equipment = pd.read_csv(equipment_path)

flight.drop(columns=['nr_linha'], inplace=True)

# save to csv with separator ;
airport.to_csv('dataset/final_data/airports.csv', sep=';', index=False)
airline.to_csv('dataset/final_data/company.csv', sep=';', index=False)
route.to_csv('dataset/final_data/line.csv', sep=';', index=False)
flight.to_csv('dataset/final_data/flights.csv', sep=';', index=False)
equipment.to_csv('dataset/final_data/equipment.csv', sep=';', index=False)
