==> data_cleaning/05_reformat_csv.ipynb <==
# Fix CSV formatting
*   Author(s): Tomaso Stefanizzi
*   Description: This notebook aims to find a workaround to the file export from DBeaver.

## Imports

import json
import csv
import pandas as pd
import numpy as np

open the JSON output from DBeaver

#read the json file
with open('/dataset/jj.json') as f:
    data = json.load(f)

open the csv output from DBeaver

with open('/dataset/output.csv') as f:
    data1 = f.read()
    print(data1.count('"\n'))

### Remove the \r from the JSON

import json

with open('/dataset/jj.json', 'r') as f:
    data = json.load(f)

results = data[list(data.keys())[0]]
for record in results[:3]:
    print(record)
cleaned_results = []

for record in results:
    cleaned_record = {}
    for key, value in record.items():
        if isinstance(value, str):
            cleaned_value = value.replace('\r', '')
            cleaned_record[key] = cleaned_value
        else:
            cleaned_record[key] = value
    cleaned_results.append(cleaned_record)
for record in cleaned_results[:3]:
    print(record)

df = pd.DataFrame(cleaned_results)
df.head()

## save into file

df.to_csv('/dataset/final_data/full_dataset.csv', index=False, sep=';')
