==> data_cleaning/06_cassandra_data_wrangling.ipynb <==
# Cassandra data wrangling

*   Author(s): Tomaso Stefanizzi
*   Description: This notebooks aims to solve some problems before inserting the data into the Cassandra database. Cassandra doesn't support all kinds of datatypes from sql, hence some adjustments has to be made.

## Imports

import pandas as pd
import numpy as np

## Import the full table

path = 'dataset/final.csv'
df = pd.read_csv(path, sep='$')
df.shape

the shape of the database is consistent with the previous flight table

## Data Wrangling

### Data Inspection

df.head()

df.columns

### Fix Time format
In Cassandra, the Time format is expressed in both int or string. In this case, I will split time into two columns, one for hours and one for seconds

df[['hour_partida_real', 'minutes_partida_real', 'seconds']] = df['hr_partida_real'].str.split(':', expand=True)
df['hour_partida_real'] = df['hour_partida_real'].astype(int)
df['minutes_partida_real'] = df['minutes_partida_real'].astype(int)
df.drop(columns=['hr_partida_real', 'seconds'], inplace=True)

df[['hour_chegada_real', 'minutes_chegada_real','seconds']] = df['hr_chegada_real'].str.split(':', expand=True)
df['hour_chegada_real'] = df['hour_chegada_real'].astype(int)
df['minutes_chegada_real'] = df['minutes_chegada_real'].astype(int)
df.drop(columns=['hr_chegada_real', 'seconds'], inplace=True)




### Fixing datetime
For the same reasons as before, I'm splitting the datetime into day, month, year columns.

df.dtypes

df['dt_partida_real'] = pd.to_datetime(df['dt_partida_real'])
df['dt_chegada_real'] = pd.to_datetime(df['dt_chegada_real'])

df['year_partida_real'] = df['dt_partida_real'].dt.year
df['month_partida_real'] = df['dt_partida_real'].dt.month
df['day_partida_real'] = df['dt_partida_real'].dt.day

df['year_chegada_real'] = df['dt_chegada_real'].dt.year
df['month_chegada_real'] = df['dt_chegada_real'].dt.month
df['day_chegada_real'] = df['dt_chegada_real'].dt.day

df.drop(columns=['dt_partida_real', 'dt_chegada_real'], inplace=True)

df.head()

seems ok

df.dtypes

## Save into file

df.to_csv('dataset/final_data/full_dataset.csv', index=False, sep=';')

dfs = pd.read_csv('dataset/final_data/full_dataset.csv', sep=';')
dfs.head()
